Refactor Team kernels.
